cleaning annoying warnings
Former-commit-id: 3c0388e8ee616b5aa89bc41512b68a5faf8f3069

+++ Client/src/ninjam/protocol/ClientMessages.cpp
@@ -257,7 +257,7 @@ void ClientIntervalUploadWrite::serializeTo(QByteArray &buffer){
     stream << intervalCompleted;
     stream.writeRawData( encodedAudioBuffer.data(), encodedAudioBuffer.size() );
 
-    assert(buffer.size() == payload + 5);
+    assert(buffer.size() == (int)(payload + 5));
 }
